move encodeExtractions into pipeline builder, delete old file
